requirements.txt
pinata
web3
python-dotenv
wave
openai
requests
playsound
pydub
sounddevice
numpy
nltk
speechrecognition
scipy
serial
mutagen
regex_spm